PolyCommitment: move chunked.rs into commitment
Defining it just below PolyComm definition.

It is mostly method definitions specialized for a particular instantiation of
the type PolyComm. We could even merge both into one. Leave this for future work.
